requirements.txt
git+https://github.com/Rapptz/discord.py
sympy
pymongo
similar
dnspython
wolframalpha
googlesearch-python
wikipedia
googletrans==3.1.0a0
nltk
textblob
newspaper3k